Rename package to not conflict with imports of java.util
