+++ community/cypher/cypher-compiler-3.2/src/main/java/org/neo4j/cypher/internal/compiler/v3_2/common/CypherOrderability.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.compiler.v3_2.util;
+package org.neo4j.cypher.internal.compiler.v3_2.common;
 
 import java.util.Arrays;
 import java.util.Comparator;
