+++ enterprise/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/spi/v3_2/codegen/AuxGenerator.scala
@@ -26,7 +26,7 @@ import org.neo4j.cypher.internal.codegen.CompiledEquivalenceUtils
 import org.neo4j.cypher.internal.compiled_runtime.v3_2.codegen.CodeGenContext
 import org.neo4j.cypher.internal.compiled_runtime.v3_2.codegen.ir.expressions.{CodeGenType, ReferenceType, RepresentationType}
 import org.neo4j.cypher.internal.compiled_runtime.v3_2.codegen.spi._
-import org.neo4j.cypher.internal.compiler.v3_2.util.CypherOrderability
+import org.neo4j.cypher.internal.compiler.v3_2.common.CypherOrderability
 import org.neo4j.cypher.internal.frontend.v3_2.helpers._
 import org.neo4j.cypher.internal.frontend.v3_2.symbols
